Move defaultConfig to its own file. Fix generation of non firestore project
