+++ .babelrc
@@ -4,7 +4,7 @@
       "@babel/preset-env",
       {
         "targets": {
-          "node": "6"
+          "node": "8"
         }
       }
     ]
