+++ examples/react-firebase-redux/src/containers/App/App.js
@@ -1,37 +1,26 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { BrowserRouter as Router } from 'react-router-dom'
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
+import { Provider } from 'react-redux'
+import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
 import firebase from 'firebase/app'
 import 'firebase/auth'
 import 'firebase/database'
-import 'firebase/firestore' // make sure you add this for firestore
-import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
-import { createFirestoreInstance } from 'redux-firestore'
-import { Provider } from 'react-redux'
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import ThemeSettings from 'theme'
+import { defaultRRFConfig } from 'defaultConfig'
 import { firebase as fbConfig, reduxFirebase as envRfConfig } from 'config'
 
 const theme = createMuiTheme(ThemeSettings)
 
-// ======================================================
-// Redux + Firebase Config (react-redux-firebase & redux-firestore)
-// ======================================================
-const defaultRRFConfig = {
-  userProfile: 'users', // root that user profiles are written to
-  updateProfileOnLogin: false, // enable/disable updating of profile on login
-  presence: 'presence', // list currently online users under "presence" path in RTDB
-  sessions: null, // Skip storing of sessions
-  enableLogging: false // enable/disable Firebase Database Logging
-  // profileDecorator: (userData) => ({ email: userData.email }) // customize format of user profile
-}
+// Initialize Firebase instance
+firebase.initializeApp(fbConfig)
+
+// Combine default and environment specific configs for react-redux-firebase
 const rrfConfig = {
   ...defaultRRFConfig,
   ...envRfConfig
 }
-
-// Initialize Firebase instance
-firebase.initializeApp(fbConfig)
 
 function App({ routes, store }) {
   return (
@@ -40,8 +29,7 @@ function App({ routes, store }) {
         <ReactReduxFirebaseProvider
           firebase={firebase}
           config={rrfConfig}
-          dispatch={store.dispatch}
-          createFirestoreInstance={createFirestoreInstance}>
+          dispatch={store.dispatch}>
           <Router>{routes}</Router>
         </ReactReduxFirebaseProvider>
       </Provider>
