--- examples/react-firebase-redux/src/modules/notification/components/withNotifications.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux'
-import * as actions from '../actions'
-
-// Wrap actions in dispatch
-export default connect(
-  null,
-  actions
-)
