+++ examples/react-firebase-redux/src/routes/Projects/routes/Project/components/ProjectPage/ProjectPage.enhancer.js
@@ -1,6 +1,3 @@
-import { compose } from 'redux'
-import { withRouter } from 'react-router-dom'
-import { setDisplayName } from 'recompose'
 import { UserIsAuthenticated } from 'utils/router'
 
 // Redirect to /login if user is not logged in
