+++ examples/redux-firestore/src/containers/App/App.js
@@ -1,51 +1,31 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { BrowserRouter as Router } from 'react-router-dom'
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
+import { Provider } from 'react-redux'
+import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
+import { createFirestoreInstance } from 'redux-firestore'
 import firebase from 'firebase/app'
 import 'firebase/auth'
 import 'firebase/database'
-import 'firebase/firestore' // make sure you add this for firestore
-import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
-import { createFirestoreInstance } from 'redux-firestore'
-import { Provider } from 'react-redux'
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
+import 'firebase/firestore'
 import { initializeMessaging } from 'utils/firebaseMessaging'
 import { setAnalyticsUser } from 'utils/analytics'
-import { setErrorUser } from '../utils/errorHandler'
+import { setErrorUser } from 'utils/errorHandler'
 import ThemeSettings from 'theme'
+import { defaultRRFConfig } from 'defaultConfig'
 import { firebase as fbConfig, reduxFirebase as envRfConfig } from 'config'
 
 const theme = createMuiTheme(ThemeSettings)
 
-// ======================================================
-// Redux + Firebase Config (react-redux-firebase & redux-firestore)
-// ======================================================
-const defaultRRFConfig = {
-  userProfile: 'users', // root that user profiles are written to
-  updateProfileOnLogin: false, // enable/disable updating of profile on login
-  presence: 'presence', // list currently online users under "presence" path in RTDB
-  sessions: null, // Skip storing of sessions
-  enableLogging: false, // enable/disable Firebase Database Logging
-  useFirestoreForProfile: true, // Save profile to Firestore instead of Real Time Database
-  useFirestoreForStorageMeta: true, // Metadata associated with storage file uploads goes to Firestore
-  onAuthStateChanged: (auth, firebaseInstance, dispatch) => {
-    if (auth) {
-      // Set auth within error handler
-      setErrorUser(auth)
-      // Initalize messaging with dispatch
-      initializeMessaging(dispatch)
-      // Set auth within analytics
-      setAnalyticsUser(auth)
-    }
-  }
-}
+// Initialize Firebase instance
+firebase.initializeApp(fbConfig)
+
+// Combine default and environment specific configs for react-redux-firebase
 const rrfConfig = {
   ...defaultRRFConfig,
   ...envRfConfig
 }
-
-// Initialize Firebase instance
-firebase.initializeApp(fbConfig)
 
 function App({ routes, store }) {
   return (
