+++ examples/redux-firestore/src/defaultConfig.js
@@ -1,0 +1,23 @@
+// ======================================================
+// Default Redux + Firebase Config (react-redux-firebase & redux-firestore)
+// Note: Differs from src/config.js which is environment specific config
+// ======================================================
+export const defaultRRFConfig = {
+  userProfile: 'users', // root that user profiles are written to
+  updateProfileOnLogin: false, // enable/disable updating of profile on login
+  presence: 'presence', // list currently online users under "presence" path in RTDB
+  sessions: null, // Skip storing of sessions
+  enableLogging: false, // enable/disable Firebase Database Logging
+  useFirestoreForProfile: true, // Save profile to Firestore instead of Real Time Database
+  useFirestoreForStorageMeta: true, // Metadata associated with storage file uploads goes to Firestore
+  onAuthStateChanged: (auth, firebaseInstance, dispatch) => {
+    if (auth) {
+      // Set auth within error handler
+      setErrorUser(auth)
+      // Initalize messaging with dispatch
+      initializeMessaging(dispatch)
+      // Set auth within analytics
+      setAnalyticsUser(auth)
+    }
+  }
+}
