+++ generators/app/templates/src/modules/notification/useNotifications.js
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 import { bindActionCreators } from 'redux'
 import { useDispatch } from 'react-redux'
-import * as actions from '../actions'
+import * as actions from './actions'
 
 /**
  * React hook for access to notifications. Returns
